move font loading to top-level provider
