+++ src/components/BlogHeader/BlogHeader.css
@@ -25,7 +25,7 @@
     position: relative;
     width: 100%;
     height: 0;
-    padding-top: 62.25%;
+    padding-top: 66.25%;
     overflow: hidden;
     box-shadow: 0 10px 40px -20px rgba(0, 0, 0, 0.325);
 
@@ -35,6 +35,7 @@
   }
 
   &__imageFigure {
+    position: absolute;
     top: 0;
     left: 0;
     width: 100%;
@@ -60,8 +61,12 @@
 
   &__image {
     display: block;
+    position: absolute;
+    top: 50%;
+    left: 50%;
     width: 100%;
     height: 100%;
+    transform: translate(-50%, -50%);
     object-fit: cover;
     object-position: 50%;
   }
