+++ src/providers/index.js
@@ -1,0 +1,4 @@
+import { FontProvider, FontContext } from './FontProvider';
+import { ThemeProvider, ThemeContext } from './ThemeProvider';
+
+export { FontContext, FontProvider, ThemeContext, ThemeProvider };
